Convert tone analyzer development script to component
