+++ components/veritasai/tone/__init__.py
@@ -1,0 +1,3 @@
+from .core import analyze
+
+__all__ = ["analyze"]
